pyproject.toml
[project]
name = "productivity-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.118.0",
    "psycopg>=3.2.10",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-keycloak>=5.8.1",
    "sqlalchemy>=2.0.43",
]
